+++ ci/build.sh
@@ -114,10 +114,10 @@ case ${BUILD_TARGERT} in
 esac
 
 if [ -n "$appveyor_build_version" -a -z "$VERSION" ]; then
-    export VERSION="v1.0.1"
+    export VERSION="v1.0.2"
 fi
 if [ -z "$VERSION" ]; then
-    export VERSION="v1.0.1"
+    export VERSION="v1.0.2"
 fi
 export UPLOADTOOL_BODY="Release Calendar ${VERSION}.<br> The change see [ChangeLog.md](ChangeLog.md) or [ChangeLog_zh_CN.md](ChangeLog_zh_CN.md)"
 #export UPLOADTOOL_PR_BODY=