+++ App/Resource/Translations/CalendarApp_zh_TW.ts
@@ -50,91 +50,83 @@
         <translation>工具(&amp;T)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="59"/>
+        <location filename="../../MainWindow.ui" line="66"/>
         <source>About(&amp;A)</source>
         <translation>關於(&amp;A)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="62"/>
-        <location filename="../../MainWindow.ui" line="65"/>
-        <location filename="../../MainWindow.ui" line="68"/>
-        <location filename="../../MainWindow.cpp" line="39"/>
+        <location filename="../../MainWindow.ui" line="69"/>
+        <location filename="../../MainWindow.ui" line="72"/>
+        <location filename="../../MainWindow.ui" line="75"/>
+        <location filename="../../MainWindow.cpp" line="40"/>
         <source>About</source>
         <translation>關於</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="77"/>
+        <location filename="../../MainWindow.ui" line="84"/>
         <source>Exit(&amp;E)</source>
         <translation>退出(&amp;E)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="86"/>
+        <location filename="../../MainWindow.ui" line="93"/>
         <source>Option(&amp;O)</source>
         <translation>選項(&amp;O)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="89"/>
-        <location filename="../../MainWindow.ui" line="92"/>
-        <location filename="../../MainWindow.ui" line="95"/>
+        <location filename="../../MainWindow.ui" line="96"/>
+        <location filename="../../MainWindow.ui" line="99"/>
+        <location filename="../../MainWindow.ui" line="102"/>
         <source>Option</source>
         <translation>選項</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="104"/>
+        <location filename="../../MainWindow.ui" line="116"/>
+        <source>Default(&amp;D)</source>
+        <translation>默認(&amp;D)</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="121"/>
+        <source>Open</source>
+        <translation>打開</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="45"/>
         <source>Sink(&amp;S)</source>
         <translation>換膚(&amp;S)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="107"/>
-        <location filename="../../MainWindow.ui" line="110"/>
-        <location filename="../../MainWindow.ui" line="113"/>
-        <location filename="../../MainWindow.ui" line="116"/>
-        <source>Sink</source>
-        <translation>換膚</translation>
-    </message>
-    <message>
-        <location filename="../../MainWindow.ui" line="125"/>
+        <location filename="../../MainWindow.ui" line="111"/>
         <source>Update(&amp;U)</source>
         <translation>更新(&amp;U)</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="29"/>
+        <location filename="../../MainWindow.cpp" line="30"/>
         <source>Exit</source>
         <translation>退出</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="33"/>
-        <location filename="../../MainWindow.cpp" line="157"/>
+        <location filename="../../MainWindow.cpp" line="34"/>
+        <location filename="../../MainWindow.cpp" line="158"/>
         <source>Hide</source>
         <translation>隱藏</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="35"/>
-        <location filename="../../MainWindow.cpp" line="161"/>
-        <location filename="../../MainWindow.cpp" line="284"/>
+        <location filename="../../MainWindow.cpp" line="36"/>
+        <location filename="../../MainWindow.cpp" line="162"/>
+        <location filename="../../MainWindow.cpp" line="231"/>
         <source>Show</source>
         <translation>顯示</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.cpp" line="43"/>
+        <location filename="../../MainWindow.cpp" line="44"/>
         <source>Enable run from boot</source>
         <translation>允許開戶自啟動</translation>
-    </message>
-    <message>
-        <location filename="../../MainWindow.cpp" line="249"/>
-        <source>Open sink</source>
-        <translation>打開皮膚</translation>
-    </message>
-    <message>
-        <location filename="../../MainWindow.cpp" line="251"/>
-        <source>Style files(*.qss)</source>
-        <translation>樣式文件(*.qss)</translation>
     </message>
 </context>
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../../main.cpp" line="47"/>
+        <location filename="../../main.cpp" line="48"/>
         <source>Calendar</source>
         <translation>日歷</translation>
     </message>